Family Member Class File
The main meat of this class is just the method that calculates calories of an individual given their information.

I thought about it and it may not be necessary to even have this class. 
We can just use the method to calculate calorie count and just store that data as a number.

Since we aren't even using their names later on, there is no need to store any of the information. (Unless we plan to extend upon our application)


I converted this from my java code to javascript with this website:
https://kalkicode.com/ai/online-java-to-javascript-converter